HSEARCH-4141 Move code related to POJO identifiers to a dedicated package
It won't be just for indexing processors anymore.
