+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/session/impl/HibernateOrmSessionIndexedTypeContext.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.search.mapper.orm.session.impl;
 
-import org.hibernate.search.mapper.pojo.bridge.runtime.spi.IdentifierMapping;
+import org.hibernate.search.mapper.pojo.identity.spi.IdentifierMapping;
 
 /**
  * @param <E> The entity type mapped to the index.
