+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/identity/impl/IdentifierMappingImplementor.java
@@ -4,13 +4,13 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.bridge.runtime.impl;
+package org.hibernate.search.mapper.pojo.identity.impl;
 
 import java.util.function.Supplier;
 
 import org.hibernate.search.mapper.pojo.bridge.runtime.spi.BridgeMappingContext;
 import org.hibernate.search.mapper.pojo.bridge.runtime.spi.BridgeSessionContext;
-import org.hibernate.search.mapper.pojo.bridge.runtime.spi.IdentifierMapping;
+import org.hibernate.search.mapper.pojo.identity.spi.IdentifierMapping;
 
 /**
  * @param <I> The identifier type for the mapped entity type.
