+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/identity/spi/IdentifierMapping.java
@@ -4,7 +4,9 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.bridge.runtime.spi;
+package org.hibernate.search.mapper.pojo.identity.spi;
+
+import org.hibernate.search.mapper.pojo.bridge.runtime.spi.BridgeSessionContext;
 
 public interface IdentifierMapping {
