+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/search/loading/impl/PojoSearchLoadingIndexedTypeContext.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.search.mapper.pojo.search.loading.impl;
 
-import org.hibernate.search.mapper.pojo.bridge.runtime.spi.IdentifierMapping;
+import org.hibernate.search.mapper.pojo.identity.spi.IdentifierMapping;
 import org.hibernate.search.mapper.pojo.loading.spi.PojoLoadingTypeContext;
 
 public interface PojoSearchLoadingIndexedTypeContext<E> extends PojoLoadingTypeContext<E> {
